+++ k9mail/src/main/AndroidManifest.xml
@@ -119,11 +119,6 @@
             android:label="@string/account_settings_composition_title"/>
 
         <activity
-            android:name=".activity.setup.AccountSetupOptions"
-            android:configChanges="locale"
-            android:label="@string/account_setup_options_title"/>
-
-        <activity
             android:name=".activity.setup.AccountSetupNames"
             android:configChanges="locale"
             android:label="@string/account_setup_names_title"/>
