+++ k9mail/src/main/java/com/fsck/k9/activity/setup/options/OptionsContract.java
@@ -1,0 +1,17 @@
+package com.fsck.k9.activity.setup.options;
+
+
+import com.fsck.k9.BasePresenter;
+import com.fsck.k9.BaseView;
+
+class OptionsContract {
+    interface View extends BaseView<Presenter> {
+        void next();
+    }
+
+    interface Presenter extends BasePresenter {
+        void onNextButtonClicked(boolean isNotifyViewChecked, boolean isNotifySyncViewClicked,
+                                 int checkFrequencyViewSelectedValue, int displayCountViewSelectedValue,
+                                 boolean isPushEnableClicked);
+    }
+}
